neo4j_microservice/requirements.txt
# Dependencies for Odoo Module Graph to Neo4j Microservice

# HTTP Requests
requests==2.31.0

# Neo4j Driver
neo4j==5.14.0

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
responses==0.24.1

# Type Hints
typing-extensions==4.8.0